+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Unreleased
 ==========
 
 - Prevent hypothetical ``ImportError`` if a Python installation is missing the built-in ``distutils`` module (`#118 <https://github.com/Delgan/loguru/issues/118>`_).
+- Fix an error using a ``filter`` function "by name" while receiving a log with ``record["name"]`` equals to ``None``.
 
 
 0.3.2 (2019-07-21)